add: 사이드바 로그아웃버튼 컴포넌트 추가
